backend/requirements.txt
# ============================================
# The North London Planning Intelligence Agent
# Python Dependencies
# ============================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database & Vector Store
supabase==2.3.4
asyncpg==0.29.0
pgvector==0.2.4

# OpenAI & LLM
openai==1.12.0
tiktoken==0.5.2
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20

# Document Processing
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
unstructured[all-docs]==0.12.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==10.2.0

# Text Processing
beautifulsoup4==4.12.3
lxml==5.1.0
html2text==2024.2.26
markdown==3.5.2
regex==2023.12.25

# Hybrid Search & Reranking
rank-bm25==0.2.2
sentence-transformers==2.3.1

# Caching
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
orjson==3.9.12

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# Development
black==24.1.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
